+++ cmd2/cmd2.py
@@ -41,10 +41,7 @@ import sys
 import tempfile
 import threading
 from code import InteractiveConsole
-from collections import (
-    OrderedDict,
-    namedtuple,
-)
+from collections import namedtuple
 from collections.abc import (
     Callable,
     Iterable,
@@ -2951,7 +2948,7 @@ class Cmd:
         # Resolve the arguments in reverse and read their values from statement.argv since those
         # are unquoted. Macro args should have been quoted when the macro was created.
         resolved = macro.value
-        reverse_arg_list = sorted(macro.arg_list, key=lambda ma: ma.start_index, reverse=True)
+        reverse_arg_list = sorted(macro.args, key=lambda ma: ma.start_index, reverse=True)
 
         for macro_arg in reverse_arg_list:
             if macro_arg.is_escaped:
@@ -3743,7 +3740,7 @@ class Cmd:
             value += ' ' + ' '.join(args.command_args)
 
         # Find all normal arguments
-        arg_list = []
+        macro_args = []
         normal_matches = re.finditer(MacroArg.macro_normal_arg_pattern, value)
         max_arg_num = 0
         arg_nums = set()
@@ -3762,7 +3759,7 @@ class Cmd:
                 arg_nums.add(cur_num)
                 max_arg_num = max(max_arg_num, cur_num)
 
-                arg_list.append(MacroArg(start_index=cur_match.start(), number_str=cur_num_str, is_escaped=False))
+                macro_args.append(MacroArg(start_index=cur_match.start(), number_str=cur_num_str, is_escaped=False))
         except StopIteration:
             pass
 
@@ -3781,7 +3778,7 @@ class Cmd:
                 # Get the number string between the braces
                 cur_num_str = re.findall(MacroArg.digit_pattern, cur_match.group())[0]
 
-                arg_list.append(MacroArg(start_index=cur_match.start(), number_str=cur_num_str, is_escaped=True))
+                macro_args.append(MacroArg(start_index=cur_match.start(), number_str=cur_num_str, is_escaped=True))
         except StopIteration:
             pass
 
@@ -3789,7 +3786,7 @@ class Cmd:
         result = "overwritten" if args.name in self.macros else "created"
         self.poutput(f"Macro '{args.name}' {result}")
 
-        self.macros[args.name] = Macro(name=args.name, value=value, minimum_arg_count=max_arg_num, arg_list=arg_list)
+        self.macros[args.name] = Macro(name=args.name, value=value, minimum_arg_count=max_arg_num, args=macro_args)
         self.last_result = True
 
     # macro -> delete
@@ -4961,7 +4958,7 @@ class Cmd:
             self.last_result = history
         return None
 
-    def _get_history(self, args: argparse.Namespace) -> 'OrderedDict[int, HistoryItem]':
+    def _get_history(self, args: argparse.Namespace) -> dict[int, HistoryItem]:
         """If an argument was supplied, then retrieve partial contents of the history; otherwise retrieve entire history.
 
         This function returns a dictionary with history items keyed by their 1-based index in ascending order.
@@ -4969,7 +4966,7 @@ class Cmd:
         if args.arg:
             try:
                 int_arg = int(args.arg)
-                return OrderedDict({int_arg: self.history.get(int_arg)})
+                return {int_arg: self.history.get(int_arg)}
             except ValueError:
                 pass