+++ cmd2/history.py
@@ -2,7 +2,6 @@
 
 import json
 import re
-from collections import OrderedDict
 from collections.abc import (
     Callable,
     Iterable,
@@ -224,7 +223,7 @@ class History(list[HistoryItem]):
     #
     spanpattern = re.compile(r'^\s*(?P<start>-?[1-9]\d*)?(?P<separator>:|(\.{2,}))(?P<end>-?[1-9]\d*)?\s*$')
 
-    def span(self, span: str, include_persisted: bool = False) -> 'OrderedDict[int, HistoryItem]':
+    def span(self, span: str, include_persisted: bool = False) -> dict[int, 'HistoryItem']:
         """Return a slice of the History list.
 
         :param span: string containing an index or a slice
@@ -273,7 +272,7 @@ class History(list[HistoryItem]):
 
         return self._build_result_dictionary(start, end)
 
-    def str_search(self, search: str, include_persisted: bool = False) -> 'OrderedDict[int, HistoryItem]':
+    def str_search(self, search: str, include_persisted: bool = False) -> dict[int, 'HistoryItem']:
         """Find history items which contain a given string.
 
         :param search: the string to search for
@@ -292,7 +291,7 @@ class History(list[HistoryItem]):
         start = 0 if include_persisted else self.session_start_index
         return self._build_result_dictionary(start, len(self), isin)
 
-    def regex_search(self, regex: str, include_persisted: bool = False) -> 'OrderedDict[int, HistoryItem]':
+    def regex_search(self, regex: str, include_persisted: bool = False) -> dict[int, 'HistoryItem']:
         """Find history items which match a given regular expression.
 
         :param regex: the regular expression to search for.
@@ -328,13 +327,13 @@ class History(list[HistoryItem]):
 
     def _build_result_dictionary(
         self, start: int, end: int, filter_func: Callable[[HistoryItem], bool] | None = None
-    ) -> 'OrderedDict[int, HistoryItem]':
+    ) -> dict[int, 'HistoryItem']:
         """Build history search results.
 
         :param start: start index to search from
         :param end: end index to stop searching (exclusive).
         """
-        results: OrderedDict[int, HistoryItem] = OrderedDict()
+        results: dict[int, HistoryItem] = {}
         for index in range(start, end):
             if filter_func is None or filter_func(self[index]):
                 results[index + 1] = self[index]