+++ CHANGELOG.md
@@ -47,6 +47,7 @@ shell, and the option for a persistent bottom bar that can display realtime stat
     - Removed `Statement.pipe_to` since it can be handled by `Statement.redirector` and
       `Statement.redirect_to`.
     - Changed `StatementParser.parse_command_only()` to return a `PartialStatement` object.
+    - Renamed `Macro.arg_list` to `Macro.args`.
 - Enhancements
     - New `cmd2.Cmd` parameters
         - **auto_suggest**: (boolean) if `True`, provide fish shell style auto-suggestions. These